imagem adicionada, refactor feito.
